refactor(cng): consolidate whisker-cng-protocol into whisker-plugin
Drops the separate `whisker-cng-protocol` crate and folds every type
+ trait + envelope into `whisker-plugin` directly. Reviewing the
prior split, the protocol crate had exactly two consumers —
`whisker-plugin` (re-exported everything) and `whisker-cng` (will
consume the engine-side API later). No 3rd-party plugin would have
ever imported `whisker-cng-protocol` directly because `whisker-plugin`
re-exports the full surface. Two crates were doing one crate's job.

The `whisker-cng` patch-bump argument the split was supposed to win
also doesn't hold up: `whisker-cng` depends on `whisker-plugin`, not
the other way round. `whisker-cng` patch bumps don't propagate to
3rd-party plugin crates as long as `whisker-plugin`'s API doesn't
change — which is the same invariant a split would have given us.

After:

  `whisker-plugin`               # Plugin trait + IRs + envelope + runner
        ↑
  `whisker-cng`                  # engine, depends on whisker-plugin
        ↑
  3rd-party plugin crate         # impl Plugin + run_as_subprocess

Mechanical changes:
- `crates/whisker-cng-protocol/` deleted (Cargo.toml + src/lib.rs).
- Workspace `members` + `[workspace.dependencies]` entries removed.
- `whisker-plugin/Cargo.toml` no longer depends on
  `whisker-cng-protocol`; description updated.
- `whisker-plugin/src/lib.rs` now owns every type the protocol crate
  used to: `Plugin`, `GenerateContext`, `AppMeta`, `IosProjectIr`,
  `PlistValue`, `PbxprojOp`, `AndroidProjectIr`, `AndroidManifest`,
  `MetaDataEntry`, `IntentFilter`, `IntentFilterData`, `GradleDsl`,
  `FileEntry`, `Target`, `Operation`, `MutationRecord`,
  `MutationJournal`, `PluginRequest`, `PluginResponse`.
- All 8 tests from both crates merged into one `tests` module;
  subprocess-runner tests prefixed `subprocess_` to disambiguate.
- Crate-level rustdoc rewritten to explain the three-audience
  design (1st-party / 3rd-party / engine) on one crate.

Verified:
- `cargo check --workspace` clean
- `cargo fmt -p whisker-plugin` clean
- `cargo clippy -p whisker-plugin --all-targets -- -D warnings` clean
- `cargo test -p whisker-plugin` — 8 unit + 1 compile-checked
  doctest, all pass
