--- crates/whisker-cng-protocol/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "whisker-cng-protocol"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-authors.workspace = true
-description = "Stable wire protocol + Plugin trait shared between whisker-cng and 3rd-party plugin binaries."
-
-[dependencies]
-# The Plugin trait's `Config` associated type requires Serialize +
-# DeserializeOwned, and every IR type round-trips through the
-# subprocess JSON envelope. Pull serde + serde_json transitively from
-# the workspace so 1st-party and 3rd-party plugin binaries link the
-# same version.
-serde = { workspace = true }
-serde_json = { workspace = true }
-anyhow = { workspace = true }
